perf(components): seperate color filters to improve bundling | bh
