+++ packages/components/src/styles/color-filters/filtersContrast.spec.ts
@@ -1,0 +1,5 @@
+import * as filtersContrast from './filtersContrast';
+
+it.each(Object.keys(filtersContrast))('should have correct filter for %s', (filterKey) => {
+  expect(filtersContrast[filterKey]).toMatchSnapshot();
+});
