+++ packages/components/src/styles/color-filters/filtersContrast.ts
@@ -1,0 +1,9 @@
+import { toFilter } from './toFilter';
+
+export const filterLightContrastLow = toFilter([93, 11, 36, 201, 89, 102]);
+export const filterLightContrastMedium = toFilter([65, 6, 119, 187, 90, 92]);
+export const filterLightContrastHigh = toFilter([40, 2, 686, 187, 80, 94]);
+
+export const filterDarkContrastLow = toFilter([20, 7, 421, 202, 97, 82]);
+export const filterDarkContrastMedium = toFilter([54, 4, 229, 187, 91, 84]);
+export const filterDarkContrastHigh = toFilter([68, 6, 108, 187, 104, 88]);
