+++ packages/components/src/styles/color-filters/filtersNotification.spec.ts
@@ -1,0 +1,5 @@
+import * as filtersNotification from './filtersNotification';
+
+it.each(Object.keys(filtersNotification))('should have correct filter for %s', (filterKey) => {
+  expect(filtersNotification[filterKey]).toMatchSnapshot();
+});
