+++ packages/components/src/styles/color-filters/filtersNotification.ts
@@ -1,0 +1,11 @@
+import { toFilter } from './toFilter';
+
+export const filterLightNotificationSuccess = toFilter([62, 61, 551, 86, 86, 80]);
+export const filterLightNotificationWarning = toFilter([74, 91, 343, 348, 92, 86]);
+export const filterLightNotificationError = toFilter([25, 62, 2003, 335, 100, 97]);
+export const filterLightNotificationInfo = toFilter([31, 32, 5493, 216, 90, 107]);
+
+export const filterDarkNotificationSuccess = toFilter([59, 22, 1342, 86, 96, 88]);
+export const filterDarkNotificationWarning = toFilter([72, 94, 303, 354, 89, 94]);
+export const filterDarkNotificationError = toFilter([28, 34, 3133, 333, 95, 100]);
+export const filterDarkNotificationInfo = toFilter([31, 32, 5493, 216, 90, 107]);
