+++ packages/components/src/styles/color-filters/filtersPrimary.spec.ts
@@ -1,0 +1,5 @@
+import * as filtersPrimary from './filtersPrimary';
+
+it.each(Object.keys(filtersPrimary))('should have correct filter for %s', (filterKey) => {
+  expect(filtersPrimary[filterKey]).toMatchSnapshot();
+});
