+++ packages/components/src/styles/color-filters/filtersPrimary.ts
@@ -1,0 +1,5 @@
+import { toFilter } from './toFilter';
+
+export const filterLightPrimary = toFilter([3, 7, 2930, 188, 91, 103]);
+
+export const filterDarkPrimary = toFilter([97, 55, 2840, 180, 114, 103]);
