+++ packages/components/src/styles/color-filters/index.ts
@@ -1,0 +1,3 @@
+export * from './filtersContrast';
+export * from './filtersNotification';
+export * from './filtersPrimary';
